Initial clean-up of old vehicle code.
Early code was written before we had stable vector code and before the Git repo;
I've removed some of the experimental stuff that was superceded by better code
later. Some of the early stuff had some good ideas that I haven't had a chance to
fully incorporate; the old_unsorted folder holds this.
