+++ vehicle/old_unsorted/README.txt
@@ -1,0 +1,4 @@
+old_unsorted directory:
+
+Modules in here were from early experimental code. They are currently not working as written,
+but contain some ideas that should be incorporated elsewhere.
